requirements.txt
aiogram==3.5.0
python-dotenv==1.0.0
ffmpeg-python==0.2.0